+++ lib/ts/recipe/passwordless/components/features/mfa/index.tsx
@@ -172,7 +172,7 @@ export function useChildProps(
             onFetchError: async (err: Response) => {
                 if (err.status === Session.getInstanceOrThrow().config.invalidClaimStatusCode) {
                     const invalidClaims = await getInvalidClaimsFromResponse({ response: err, userContext });
-                    if (invalidClaims.some((i) => i.validatorId === EmailVerificationClaim.id)) {
+                    if (invalidClaims.some((i) => i.id === EmailVerificationClaim.id)) {
                         try {
                             // it's OK if this throws,
                             const evInstance = EmailVerification.getInstanceOrThrow();
@@ -362,7 +362,7 @@ function useOnLoad(
                             err.status === SessionRecipe.getInstanceOrThrow().config.invalidClaimStatusCode
                         ) {
                             const invalidClaims = await getInvalidClaimsFromResponse({ response: err, userContext });
-                            if (invalidClaims.some((i) => i.validatorId === EmailVerificationClaim.id)) {
+                            if (invalidClaims.some((i) => i.id === EmailVerificationClaim.id)) {
                                 try {
                                     // it's OK if this throws,
                                     const evInstance = EmailVerificationRecipe.getInstanceOrThrow();