+++ lib/ts/recipe/thirdparty/components/features/signInAndUpCallback/index.tsx
@@ -122,7 +122,7 @@ const SignInAndUpCallback: React.FC<PropType> = (props) => {
         async (err: any) => {
             if ("status" in err && err.status === Session.getInstanceOrThrow().config.invalidClaimStatusCode) {
                 const invalidClaims = await getInvalidClaimsFromResponse({ response: err, userContext });
-                if (invalidClaims.some((i) => i.validatorId === EmailVerificationClaim.id)) {
+                if (invalidClaims.some((i) => i.id === EmailVerificationClaim.id)) {
                     try {
                         // it's OK if this throws,
                         const evInstance = EmailVerification.getInstanceOrThrow();